+++ src/__tests__/utils-test.ts
@@ -6,11 +6,13 @@ import {
     share,
     isABTestingAvailable,
     reportStatus,
+    fetch,
 } from '../utils';
 import {
     createFakeAndroidPostMessage,
     removeFakeAndroidPostMessage,
 } from './fake-post-message';
+import {isWebViewBridgeAvailable} from '../post-message';
 
 const ANY_STRING = 'any-string';
 
@@ -71,7 +73,7 @@ test('share', cb => {
 test('set webview title', cb => {
     createFakeAndroidPostMessage({
         checkMessage: message => {
-            expect(message.type).toBe('SET_TITLE');
+            expect(message.type).toBe('NAVIGATION_BAR');
             expect(message.payload).toEqual({title: ANY_STRING});
         },
         getResponse: message => ({
@@ -184,12 +186,11 @@ test('notify page loaded', cb => {
     });
 });
 
-test('request remote configuration', cb => {
+test('isABTestingAvailable happy case', async () => {
     const REMOTE_CONFIGURATION = {
         result: {
             key1: 'true',
             key2: 'false',
-            key3: 'true',
         },
     };
 
@@ -205,13 +206,22 @@ test('request remote configuration', cb => {
         }),
     });
 
-    isABTestingAvailable('key1').then(res => {
+    await isABTestingAvailable('key1').then(res => {
         expect(res).toEqual(true);
-        cb();
     });
-    isABTestingAvailable('key2').then(res => {
+    await isABTestingAvailable('key2').then(res => {
         expect(res).toEqual(false);
-        cb();
+    });
+});
+
+test('isABTestingAvailable without bridge', async () => {
+    expect(isWebViewBridgeAvailable()).toBe(false);
+
+    await isABTestingAvailable('key3').then(res => {
+        expect(res).toEqual(false);
+    });
+    await isABTestingAvailable('key4').then(res => {
+        expect(res).toEqual(false);
     });
 });
 
@@ -257,5 +267,80 @@ test('report account status', cb => {
     }).then(res => {
         expect(res).toBeUndefined();
         cb();
+    });
+});
+
+test('fetch happy case', async () => {
+    const request = {
+        url: 'https://example.com',
+        method: 'GET' as 'GET',
+        headers: {key1: 'value1', key2: 'value2'},
+        body: 'hello',
+    };
+
+    const response = {
+        status: 200,
+        headers: {key3: 'value3'},
+        body: 'bye',
+    };
+
+    createFakeAndroidPostMessage({
+        checkMessage: message => {
+            expect(message.type).toBe('FETCH');
+            expect(message.payload).toEqual(request);
+        },
+        getResponse: message => ({
+            type: message.type,
+            id: message.id,
+            payload: response,
+        }),
+    });
+
+    await fetch(request).then(res => {
+        expect(res).toEqual(response);
+    });
+});
+
+test('fetch without bridge', async () => {
+    await fetch({
+        url: 'https://example.com',
+        method: 'GET' as 'GET',
+        headers: {key1: 'value1', key2: 'value2'},
+        body: 'hello',
+    }).then(res => {
+        expect(res).toEqual({
+            status: 500,
+            headers: {},
+            body: 'Bridge not available',
+        });
+    });
+});
+
+test('fetch call failure', async () => {
+    const request = {
+        url: 'https://example.com',
+        method: 'GET' as 'GET',
+        headers: {key1: 'value1', key2: 'value2'},
+        body: 'hello',
+    };
+
+    createFakeAndroidPostMessage({
+        checkMessage: message => {
+            expect(message.type).toBe('FETCH');
+            expect(message.payload).toEqual(request);
+        },
+        getResponse: message => ({
+            type: 'ERROR',
+            id: message.id,
+            payload: {reason: 'Unknown'},
+        }),
+    });
+
+    await fetch(request).then(res => {
+        expect(res).toEqual({
+            status: 500,
+            headers: {},
+            body: 'Bridge call failed',
+        });
     });
 });