+++ src/post-message.ts
@@ -1,6 +1,19 @@
 import {getId} from './message-id';
 
-type MessageTypeToResponseMap = {
+/**
+ * This type map represents all the possible messages we can receive from native app
+ * Those messages can be:
+ *     - Requests: native app initiates the dialog
+ *     - Responses: native app responds to a request initiated by the web
+ */
+export type IncomingMessageMap = {
+    // Requests
+    EVENT: {
+        type: 'NATIVE_EVENT';
+        id: string;
+        payload: {event: string};
+    };
+    // Responses
     SIM_ICC: {
         id: string;
         type: 'SIM_ICC';
@@ -91,9 +104,18 @@ type MessageTypeToResponseMap = {
         type: 'STATUS_REPORT';
         payload: void;
     };
+    FETCH: {
+        id: string;
+        type: 'FETCH';
+        payload: {
+            status: number;
+            headers: {[key: string]: string};
+            body: string;
+        };
+    };
 };
 
-type Response = MessageTypeToResponseMap[keyof MessageTypeToResponseMap];
+type Response = IncomingMessageMap[keyof IncomingMessageMap];
 type Listener = (message: Response) => void;
 
 const BRIDGE = '__tuenti_webview_bridge';
@@ -149,15 +171,14 @@ export const isWebViewBridgeAvailable = (): boolean =>
     hasAndroidPostMessage() || hasWebKitPostMessage();
 
 /**
- * Send message to native app
+ * Send message to native app and waits for response
  */
-export const postMessageToNativeApp = <
-    T extends keyof MessageTypeToResponseMap
->(
+export const postMessageToNativeApp = <T extends keyof IncomingMessageMap>(
     {type, id = getId(), payload}: {type: T; id?: string; payload?: Object},
     timeout?: number,
-): Promise<MessageTypeToResponseMap[T]['payload']> => {
+): Promise<IncomingMessageMap[T]['payload']> => {
     const postMessage = getWebViewPostMessage();
+    const message = JSON.stringify({type, id, payload});
 
     if (!postMessage) {
         return Promise.reject({
@@ -165,8 +186,6 @@ export const postMessageToNativeApp = <
             reason: 'WebView postMessage not available',
         });
     }
-
-    const message = JSON.stringify({type, id, payload});
 
     // ensure postMessage call is async
     setTimeout(() => {
@@ -176,7 +195,7 @@ export const postMessageToNativeApp = <
     return new Promise((resolve, reject) => {
         let timedOut = false;
 
-        const listener = (response: MessageTypeToResponseMap[T]): void => {
+        const listener: Listener = response => {
             if (response.id === id && !timedOut) {
                 if (response.type === type) {
                     resolve(response.payload);
@@ -217,4 +236,38 @@ window[BRIDGE] = window[BRIDGE] || {
         }
         messageListeners.forEach(f => f(message));
     },
+};
+
+export type NativeEventHandler = ({
+    event,
+}: {
+    event: string;
+}) => {action: 'default'};
+
+export const onNativeEvent = (handler: NativeEventHandler) => {
+    const listener: Listener = message => {
+        if (message.type === 'NATIVE_EVENT') {
+            const response = handler({
+                event: message.payload.event,
+            });
+
+            const postMessage = getWebViewPostMessage();
+            if (postMessage) {
+                postMessage(
+                    JSON.stringify({
+                        type: 'NATIVE_EVENT',
+                        id: message.id,
+                        payload: {
+                            action: response.action || 'default',
+                        },
+                    }),
+                );
+            }
+        }
+    };
+    subscribe(listener);
+
+    return () => {
+        unsubscribe(listener);
+    };
 };